+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # 4.1.0 (not released yet)
 - Feature: Option to overwrite the global configuration per script
+- Feature: Reload from file with "=" to set absolute index
 - Fix: Add missing variables and links to README
 
 # 4.0.1